Add argument parser module for stock processing
